changed weather folder and namespace, fixed small issues, introduced configurable weather forecast
