--- Kachelmann/Sunshine.cs
@@ -1,12 +1,0 @@
-﻿namespace RaspPiTest.Kachelmann
-{
-    public enum Sunshine
-    {
-        Wolkenlos,
-        Bedeckt,
-        StarkBewoelkt,
-        LeichtBewoelkt,
-        Regen,
-        Gewitter
-    }
-}
