--- Kachelmann/Wind.cs
@@ -1,9 +1,0 @@
-﻿namespace RaspPiTest.Kachelmann
-{
-    public class Wind
-    {
-        public WindDirection Direction { get; set; }
-        public float Speed { get; set; }
-        public float MaxSpeed { get; set; }
-    }
-}
