--- Kachelmann/WindDirection.cs
@@ -1,23 +1,0 @@
-﻿namespace RaspPiTest.Kachelmann
-{
-    public enum WindDirection
-    {
-        None,
-        North,
-        NortNorthEast,
-        NorthEast,
-        EastNorthEast,
-        East,
-        EastSouthEast,
-        SouthEast,
-        SouthSouthEast,
-        South,
-        SouthSouthWest,
-        SouthWest,
-        WestSouthWest,
-        West,
-        WestNorthWest,
-        NorthWest,
-        NorthNorthWest
-    }
-}
