+++ Weather/Sunshine.cs
@@ -1,0 +1,12 @@
+﻿namespace RaspPiTest.Weather
+{
+    public enum Sunshine
+    {
+        Wolkenlos,
+        Bedeckt,
+        StarkBewoelkt,
+        LeichtBewoelkt,
+        Regen,
+        Gewitter
+    }
+}
