+++ Weather/WeatherFetch.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace RaspPiTest.Weather
+{
+    public class WeatherFetch
+    {
+        public DateTime Fetched { get; set; }
+        public DayForecast Today { get; set; }
+        public DayForecast Tommorrow { get; set; }
+        public DayForecast DayAfterTommorrow { get; set; }
+    }
+}
