+++ Weather/WeatherOptions.cs
@@ -1,0 +1,8 @@
+﻿namespace RaspPiTest.Weather
+{
+    public class WeatherOptions
+    {
+        public int CityId { get; set; } = 2929670;
+        public int RefreshInterval { get; set; } = 120;
+    }
+}
