+++ Weather/Wind.cs
@@ -1,0 +1,9 @@
+﻿namespace RaspPiTest.Weather
+{
+    public class Wind
+    {
+        public WindDirection Direction { get; set; }
+        public float Speed { get; set; }
+        public float MaxSpeed { get; set; }
+    }
+}
