+++ Weather/WindDirection.cs
@@ -1,0 +1,23 @@
+﻿namespace RaspPiTest.Weather
+{
+    public enum WindDirection
+    {
+        None,
+        North,
+        NortNorthEast,
+        NorthEast,
+        EastNorthEast,
+        East,
+        EastSouthEast,
+        SouthEast,
+        SouthSouthEast,
+        South,
+        SouthSouthWest,
+        SouthWest,
+        WestSouthWest,
+        West,
+        WestNorthWest,
+        NorthWest,
+        NorthNorthWest
+    }
+}
