+++ appsettings.json
@@ -5,5 +5,9 @@
             "Default": "Trace"
         }
     },
-    "AllowedHosts": "*"
+    "AllowedHosts": "*",
+    "weather": {
+        "cityId": 2929670,
+        "refreshInterval":  120 
+    } 
 }
